vbpt: vbpt_cur_replace needs to check for COW
There was a bug in the previous implementation, since vbpt_cur_replace() did
not check if it needed to check for COWing the parent node. The erroneous
assumption was that since merge was run on the parent, both versions are after
vj. However, because we need to synchronize the cursors we might end up going
deeper than the nodes the previous do_merge() checked.