split ver_rebase into _{prepare,commit,abort}
If the mtree can change when performing a merge, the version we try to rebase
to might get removed from the chain. To avoid this bug, we split ver_rebase()
into to phases: prepare and commit/abort. The first phase, happens under the
lock and grabs a children reference so that the version won't get removed.